+++ jannovar-cli/src/main/java/de/charite/compbio/jannovar/cmd/annotate_pos/AnnotatePositionCommand.java
@@ -8,16 +8,17 @@ import org.apache.commons.cli.ParseException;
 import de.charite.compbio.jannovar.JannovarException;
 import de.charite.compbio.jannovar.JannovarOptions;
 import de.charite.compbio.jannovar.annotation.AllAnnotationListTextGenerator;
-import de.charite.compbio.jannovar.annotation.AnnotationList;
-import de.charite.compbio.jannovar.annotation.AnnotationListTextGenerator;
+import de.charite.compbio.jannovar.annotation.AnnotationException;
 import de.charite.compbio.jannovar.annotation.BestAnnotationListTextGenerator;
+import de.charite.compbio.jannovar.annotation.VariantAnnotations;
+import de.charite.compbio.jannovar.annotation.VariantAnnotationsTextGenerator;
 import de.charite.compbio.jannovar.annotation.VariantAnnotator;
 import de.charite.compbio.jannovar.annotation.builders.AnnotationBuilderOptions;
 import de.charite.compbio.jannovar.cmd.CommandLineParsingException;
 import de.charite.compbio.jannovar.cmd.HelpRequestedException;
 import de.charite.compbio.jannovar.cmd.JannovarAnnotationCommand;
-import de.charite.compbio.jannovar.reference.GenomeVariant;
 import de.charite.compbio.jannovar.reference.GenomePosition;
+import de.charite.compbio.jannovar.reference.GenomeVariant;
 import de.charite.compbio.jannovar.reference.PositionType;
 import de.charite.compbio.jannovar.reference.Strand;
 
@@ -58,18 +59,19 @@ public class AnnotatePositionCommand extends JannovarAnnotationCommand {
 			final GenomeVariant genomeChange = parseGenomeChange(chromosomalChange);
 
 			// Construct VariantAnnotator for building the variant annotations.
-			AnnotationList annoList = null;
+			VariantAnnotations annoList = null;
 			try {
-				annoList = annotator.buildAnnotationList(genomeChange);
+				annoList = annotator.buildAnnotations(genomeChange);
 			} catch (Exception e) {
 				System.err.println(String.format("[ERROR] Could not annotate variant %s!", chromosomalChange));
+				e.printStackTrace(System.err);
 				continue;
 			}
 
 			// Obtain first or all functional annotation(s) and effect(s).
 			final String annotation;
 			final String effect;
-			AnnotationListTextGenerator textGenerator;
+			VariantAnnotationsTextGenerator textGenerator;
 			if (options.showAll)
 				textGenerator = new AllAnnotationListTextGenerator(annoList, 0, 1);
 			else